+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added support for duplo hosts with or without scheme (http/https)
+- Added update_otherdockerconfig command in service
 
 ## [0.3.7] - 2025-09-17